+++ src/main/java/org/gdg/zipte_gdg/api/service/toss/TossService.java
@@ -1,6 +1,6 @@
 package org.gdg.zipte_gdg.api.service.toss;
 
-import org.gdg.zipte_gdg.api.controller.Toss.request.ConfirmPaymentRequestDto;
+import org.gdg.zipte_gdg.api.controller.toss.request.ConfirmPaymentRequestDto;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
