+++ src/main/java/org/gdg/zipte_gdg/api/service/toss/response/TossPaymentErrorResponse.java
@@ -1,4 +1,4 @@
-package org.gdg.zipte_gdg.domain.Toss;
+package org.gdg.zipte_gdg.api.service.toss.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class PaymentErrorResponse {
+public class TossPaymentErrorResponse {
 
     private int code;
     private String message;
