--- src/main/java/org/gdg/zipte_gdg/domain/Toss/TossPaymentMethod.java
@@ -1,7 +1,0 @@
-package org.gdg.zipte_gdg.domain.Toss;
-
-public enum TossPaymentMethod {
-
-    BILLING, CARD, BRAND
-
-}
