--- src/main/java/org/gdg/zipte_gdg/domain/Toss/TossPaymentStatus.java
@@ -1,6 +1,0 @@
-package org.gdg.zipte_gdg.domain.Toss;
-
-public enum TossPaymentStatus {
-
-    Hello,NEW, WHY
-}
