+++ src/main/java/org/gdg/zipte_gdg/domain/toss/TossPaymentRepository.java
@@ -1,4 +1,4 @@
-package org.gdg.zipte_gdg.domain.Toss;
+package org.gdg.zipte_gdg.domain.toss;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
